server/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
openai==1.0.0
httpx==0.24.1
beautifulsoup4==4.12.2
chromadb==0.4.24
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
tiktoken==0.4.0
